requirements.txt
appnope==0.1.2
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
click==8.0.3
decorator==5.1.1
Django==3.2.11
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
executing==0.8.2
Faker==11.3.0
flake8==4.0.1
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
text-unidecode==1.3
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
